CumulusTestNode for testing
