+++ test/service/Cargo.toml
@@ -7,6 +7,7 @@ edition = '2018'
 [dependencies]
 ansi_term = "0.12.1"
 codec = { package = 'parity-scale-codec', version = '1.0.0' }
+rand = "0.7.3"
 serde = { version = "1.0.101", features = ["derive"] }
 
 # Substrate
@@ -25,12 +26,15 @@ sp-blockchain = { git = "https://github.com/paritytech/substrate", branch = "roc
 sp-consensus = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 sp-inherents = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
+sp-keyring = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 sp-offchain = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "rococo-branch" }
 sp-session = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
+sp-state-machine = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 sp-timestamp = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 sp-transaction-pool = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 sp-trie = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
+substrate-test-client = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 
 # Polkadot
 polkadot-collator = { git = "https://github.com/paritytech/polkadot", branch = "rococo-branch" }
@@ -50,19 +54,15 @@ test-primitives = { path = "../primitives" }
 jsonrpc-core = "14.2.0"
 
 [dev-dependencies]
-futures = { version = "0.3.1", features = ["compat"] }
-rand = "0.7.3"
+futures = { version = "0.3.1" }
 tokio = { version = "0.2.21", features = ["macros"] }
 
 # Polkadot dependencies
-polkadot-primitives = { git = "https://github.com/paritytech/polkadot", branch = "rococo-branch" }
 polkadot-runtime-common = { git = "https://github.com/paritytech/polkadot", branch = "rococo-branch" }
 polkadot-test-runtime = { git = "https://github.com/paritytech/polkadot", branch = "rococo-branch" }
-polkadot-test-runtime-client = { git = "https://github.com/paritytech/polkadot", branch = "rococo-branch" }
 polkadot-test-service = { git = "https://github.com/paritytech/polkadot", branch = "rococo-branch" }
 
 # Substrate dependencies
 pallet-sudo = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
-substrate-test-client = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 substrate-test-runtime-client = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
 substrate-test-utils = { git = "https://github.com/paritytech/substrate", branch = "rococo-branch" }
